Create elements once rather than every render
This was a massive issue, obviously for performance, but actually it caused bigger
problems with React - since it was seeing a different element every time it destroyed and
recreated the entire DOM. This was costly and meant that transitions wouldn't work at all.

Elements are now created beforehand, and I baked some support into the themeHelper so
that styles can be added and removed via named props.